refactor: 타이머 설정 UI TimerSetting 컴포넌트로 분리
