+++ src/pages/FocusPage/Focus.jsx
@@ -7,15 +7,13 @@ import Toast from "../../components/Toast/Toast";
 import Modal from "../../components/Modal/Modal";
 import useTimer from "../../hooks/timer/useTimer";
 import { TIMER_STATUS } from "../../hooks/timer/timerConstants";
+import TimerSetting from "./components/TimerSetting";
 import ResumeConfirmPopup from "./components/ResumeConfirmPopup";
 import StopConfirmPopup from "./components/StopConfirmPopup";
 import SessionTitleInput from "./components/SessionTitleInput";
 import SessionListModal from "./components/SessionListModal";
 import formatTime from "./formatTime";
 import styles from "./Focus.module.css";
-
-// 타이머 프리셋 버튼: 버튼 클릭 시 바로 적용되는 시간(분)
-const PRESETS = [15, 25, 50];
 
 function Focus() {
   const navigate = useNavigate();
@@ -242,40 +240,13 @@ function Focus() {
           </h3>
 
           {/* 타이머 시작 전/완료: 집중 시간 설정 UI 표시 */}
-          {isIdle || isCompleted ? (
-            <div className={styles.timerPresets}>
-              {/* 프리셋 버튼: 버튼 클릭 시 15분/25분/50분 시간 설정 */}
-              {PRESETS.map((min) => (
-                <button
-                  key={min}
-                  type="button"
-                  className={`
-                    ${styles.presetBtn} 
-                    ${!isEditing && durationSec === min * 60 ? styles.active : ""}
-                  `}
-                  onClick={() => handleSelectPreset(min)}
-                >
-                  {min}분
-                </button>
-              ))}
-              {/* 직접 입력 버튼: 버튼 클릭 시 분/초 직접 설정 */}
-              <button
-                type="button"
-                className={`${styles.presetBtn} ${isEditing ? styles.active : ""}`}
-                onClick={handleSelectCustom}
-              >
-                직접 입력
-              </button>
-            </div>
-          ) : (
-            <div className={styles.timerSetTime}>
-              {/* 타이머 진행 중/일시중지: 타이머 설정 시간 표시 */}
-              <span className={`ic timer ${styles.timerSetIcon}`}></span>
-              <span className={styles.timerSetValue}>
-                {formatTime(durationSec)}
-              </span>
-            </div>
-          )}
+          <TimerSetting
+            durationSec={durationSec}
+            isEditing={isEditing}
+            isActive={isRunning || isPaused}
+            onSelectPreset={handleSelectPreset}
+            onSelectCustom={handleSelectCustom}
+          />
         </div>
 
         <div
