+++ src/pages/FocusPage/components/TimerSetting.module.css
@@ -1,0 +1,61 @@
+/* 타이머 진행 중/일시정지일 때 타이머 설정 시간 표시 */
+.timerSetTime {
+  display: flex;
+  align-items: center;
+  gap: 0.25rem;
+  position: absolute;
+  top: 100%;
+  margin-top: 1.5rem;
+  width: 5.1875rem;
+  height: 1.9375rem;
+  padding: 0.375rem 0.5rem;
+  border-radius: 3.125rem;
+  border: 1px solid rgba(0, 0, 0, 0.1);
+  background: rgba(255, 255, 255, 0.3);
+}
+
+.timerSetIcon {
+  font-size: 1.1875rem;
+}
+
+.timerSetValue {
+  font: var(--text-16-medium);
+  line-height: 1.1875rem;
+  color: var(--black);
+}
+
+/* 초기 상태/완료 상태: 타이머 시간 설정(프리셋 버튼, 직접 입력) */
+.timerPresets {
+  display: flex;
+  gap: 0.8rem;
+  margin-top: 1.5rem;
+}
+
+.presetBtn {
+  padding: 0.375rem 0.8rem;
+  border-radius: 0.7rem;
+  border: 1px solid rgba(0, 0, 0, 0.1);
+  font: var(--text-16-medium);
+  color: var(--black);
+  background: rgba(255, 255, 255, 0.3);
+  transition: all 0.15s ease;
+}
+
+.presetBtn:hover {
+  background-color: var(--green);
+}
+
+.presetBtn:active {
+  transform: scale(0.97);
+}
+
+.presetBtn:focus-visible {
+  outline: 2px solid var(--brand);
+  outline-offset: 2px;
+}
+
+.presetBtn.active {
+  background-color: var(--brand);
+  color: var(--white);
+  border-color: var(--brand);
+}
